added AlertDialog for UpdateFragment.kt.

+++ app/src/main/java/com/mtah/todolist/fragments/UpdateFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.util.Log
 import android.view.*
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
@@ -11,29 +12,25 @@ import androidx.navigation.fragment.navArgs
 import com.mtah.todolist.R
 import com.mtah.todolist.SharedViewModel
 import com.mtah.todolist.backend.ToDoViewModel
-import com.mtah.todolist.backend.models.Priority
 import com.mtah.todolist.backend.models.ToDo
-import kotlinx.android.synthetic.main.fragment_home.*
 import kotlinx.android.synthetic.main.fragment_update.*
 import kotlinx.android.synthetic.main.fragment_update.view.*
 
 class UpdateFragment : Fragment(){
 
-    val args by navArgs<UpdateFragmentArgs>()
+    private val args by navArgs<UpdateFragmentArgs>()
     private val sharedViewModel: SharedViewModel by viewModels()
     private val toDoViewModel: ToDoViewModel by viewModels()
 
-    val TAG = "UpdateFragment"
+    private val TAG = "UpdateFragment"
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-
         val view = inflater.inflate(R.layout.fragment_update, container, false)
         setHasOptionsMenu(true)
-
 
         view.title_et.setText(args.todoItem.title)
         view.description_et.setText(args.todoItem.description)
@@ -50,7 +47,7 @@ class UpdateFragment : Fragment(){
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             R.id.option_save -> updateItem()
-            R.id.option_delete -> deleteItem()
+            R.id.option_delete -> deleteConfirmation()
         }
         return super.onOptionsItemSelected(item)
     }
@@ -66,17 +63,32 @@ class UpdateFragment : Fragment(){
             val updatedItem = ToDo(args.todoItem.id, title, sharedViewModel.getPriority(priority), description)
             toDoViewModel.update(updatedItem)
             Toast.makeText(requireContext(), "Successfully changed!", Toast.LENGTH_SHORT).show()
-
+            Log.i(TAG, "updateItem: To do updated!")
             findNavController().navigate(R.id.action_updateFragment_to_homeFragment)
         } else {
             Toast.makeText(requireContext(), "Please fill out all fields!", Toast.LENGTH_SHORT).show()
         }
     }
 
-    private fun deleteItem() {
-        toDoViewModel.delete(args.todoItem)
-        Toast.makeText(requireContext(), "Deleted!", Toast.LENGTH_SHORT).show()
-        findNavController().navigate(R.id.action_updateFragment_to_homeFragment)
+    private fun deleteConfirmation(){
+        val alertBuilder = AlertDialog.Builder(requireContext())
+        alertBuilder.setMessage("Are you sure you want to delete this?")
+            .setTitle("Delete ${args.todoItem.title}")
+            .setCancelable(true)
+            .setPositiveButton("Yes") { _, _ ->
+                toDoViewModel.delete(args.todoItem)
+                Toast.makeText(
+                    requireContext(),
+                    "Deleted!",
+                    Toast.LENGTH_SHORT)
+                    .show()
+                findNavController().navigate(R.id.action_updateFragment_to_homeFragment)
+            }
+            .setNegativeButton("No") { _, _ -> }
+
+        alertBuilder.create().show()
+
+
     }
 
 }
